tech_shop/requirements.txt
asgiref==3.8.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==44.0.3
dj-database-url==3.0.1
Django==5.2
django-allauth==65.7.0
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
gunicorn==23.0.0
idna==3.10
joblib==1.5.2
numpy==2.3.3
packaging==25.0
pillow==11.3.0
psycopg2==2.9.11
psycopg2-binary==2.9.11
pycparser==2.22
PyJWT==2.9.0
python-decouple==3.8
python-dotenv==1.1.1
requests==2.32.3
scikit-learn==1.7.2
scipy==1.16.2
sqlparse==0.5.3
stripe==12.3.0
threadpoolctl==3.6.0
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.4.0
whitenoise==6.11.0
